refactoring some components
