+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Layout from './Layout';
+import Layout from './components/Layout.jsx';
 
 class App extends Component {
   render() {
