+++ src/components/ChatHeader.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { withStyles } from 'material-ui/styles';
 import AppBar from 'material-ui/AppBar';
 import Toolbar from 'material-ui/Toolbar';
@@ -11,21 +11,18 @@ const styles = theme => ({
   },
 });
 
-class ChatHeader extends Component {
-  render() {
-    const { classes } = this.props;
+const ChatHeader = props => {
+  const { classes } = props;
 
-    return (
-      <AppBar position="absolute" className={classes.appBar}>
-        <Toolbar>
-          <Typography variant="title" color="inherit" noWrap>
-            DogeCodes React Chat
-          </Typography>
-        </Toolbar>
-      </AppBar>
-    );
-  }
-
+  return (
+    <AppBar position="absolute" className={classes.appBar}>
+      <Toolbar>
+        <Typography variant="title" color="inherit" noWrap>
+          DogeCodes React Chat
+        </Typography>
+      </Toolbar>
+    </AppBar>
+  );
 }
